refactor(print): change parameter type from long to int in print function

+++ src/main.cpp
@@ -93,7 +93,7 @@ void print(const char *message)
   }
 }
 
-void print(long number, byte base = DEC)
+void print(int number, byte base = DEC)
 {
   if (!debugMode)
     return;
